% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dstm}
\alias{dstm}
\title{Fits a dynamic spatio-temporal model (DSTM)}
\usage{
dstm(Y, F_, G_0, m_0, C_0, n_samples, verbose = FALSE, sample_sigma2 = TRUE,
  discount = TRUE, sample_G = FALSE)
}
\arguments{
\item{Y}{S by T matrix containing response variable at S spatial locations and T time points}

\item{F_}{S by p matrix defining \eqn{Y_t = F \theta_t + V}}

\item{G_0}{p by p matrix defining \eqn{\theta_t = G \theta_{t-1} + W}.
If sample_G is TRUE, then this is used as the starting value and prior mean for G.}

\item{m_0}{p by 1 column vector for a priori mean of \eqn{\theta}}

\item{C_0}{p by p matrix of for a priori variance-covariance matrix of \eqn{\theta}}
}
\description{
Fits a dynamic spatio-temporal model (DSTM)
}
\examples{
# Duhh...nothing yet
}
