% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstm.R
\name{dstm_eof}
\alias{dstm_eof}
\title{Fits a dynamic spatio-temporal model with EOFs}
\usage{
dstm_eof(Y, proc_model = "RW", proc_error = "IW", P = 10L,
  n_samples = 1L, sample_sigma2 = TRUE, verbose = FALSE,
  params = NULL)
}
\arguments{
\item{Y}{(numeric matrix) S by T data matrix containing response variable
at S spatial locations and T time points}

\item{proc_model}{(character string) Process model; one of "RW" (random walk), 
"AR" (diagonal process matrix), or "Dense" (dense process matrix)}

\item{proc_error}{(character string) Process error; "IW" (inverse-Wishart) or 
"Discount" (discount factor)}

\item{P}{(integer) Number of EOFs}

\item{sample_sigma2}{(logical) whether to sample the variance of the iid
observation error}

\item{verbose}{(logical) Whether to print additional information 
(e.g., iteration in sampling algorithm)}

\item{params}{(list) List of hyperparameter values (see details)}
}
\description{
Fits a dynamic spatio-temporal model with EOFs
}
\details{
This section explains how to specify custom hyperparameters using the `params` argument.
You may specify the following as named elements of the `params` list:

-"m_0": (numeric vector) The prior mean of the state vector at time zero
 (\eqn{\theta_0})

-"C_0": (numeric matrix) The prior variance-covariance matrix of the state
 vector at time zero (\eqn{\theta_0})

-"alpha_sigma2", "beta_sigma2": The inverse-Gamma parameters (scale parameterization)
of the prior distribution on \eqn{\sigma^2}

-"alpha_lambda", "beta_lambda": The inverse-Gamma parameters (scale parameterization)
of the prior distribution on \eqn{\lambda = \delta / (1 - \delta)}
}
\examples{
# Duhh...nothing yet
}
