% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dstm_IW}
\alias{dstm_IW}
\title{Fits a DSTM using a wishart prior for W}
\usage{
dstm_IW(Y, F, G_0, Sigma_G_inv, m_0, C_0, C_W, params, proc_model,
  n_samples, verbose)
}
\description{
Fits a DSTM using a wishart prior for W
}
\examples{
# Duhh...nothing yet
}
