% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dstm_sample_G}
\alias{dstm_sample_G}
\title{Fits a dynamic spatio-temporal model (DSTM) that samples the matrix G defining the state equation}
\usage{
dstm_sample_G(Y, n_samples, p, verbose, sample_sigma2)
}
\arguments{
\item{Y}{S by T matrix containing response variable at S spatial locations and T time points}

\item{n_samples}{integer; number of posterior samples to take}

\item{p}{integer; dimension of G in the state equation \eqn{\theta_{t+1} = G \theta_{t}}}

\item{verbose}{boolean; controls verbosity}

\item{sample_sigma2}{whether boolean; to sample \eqn{\sigma^2}}
}
\description{
Fits a dynamic spatio-temporal model (DSTM) that samples the matrix G defining the state equation
}
\examples{
# Duhh...nothing yet
}
