% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FFBS}
\alias{FFBS}
\title{Performs FFBS}
\usage{
FFBS(Y, F_, V, G, W, m_0, C_0)
}
\arguments{
\item{Y}{j by T matrix containing response variable}

\item{F_}{j by p matrix defining \eqn{Y_t = F \theta_t + V}}

\item{V}{j by j variance-covariance matrix of \eqn{V}}

\item{G}{p by p matrix defining \eqn{\theta_t = G \theta_{t-1} + W}}

\item{W}{p by p variance-covariance matrix of \eqn{W}}

\item{m_0}{p by 1 column vector for a priori mean of \eqn{\theta}}

\item{C_0}{p by p matrix of for a priori variance-covariance matrix of \eqn{\theta}}
}
\description{
Performs FFBS
}
\examples{
# Duhh...nothing yet
}
