% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{ide_locations}
\alias{ide_locations}
\title{Spatial locations for IDE data sets}
\format{A numeric matrix with 400 rows and 2 columns}
\source{
Generated as part of Easton Huch's MS project at BYU
}
\usage{
ide_locations
}
\description{
A matrix containing the 400 two-dimensional spatial locations of the 
following data sets: `ide_standard` and `ide_spatially_varying`.
The rows of these two data sets correspond with the rows of `ide_locations`.
}
\seealso{
[ide_standard]/[ide_spatially_varying], the data sets that
correspond to these spatial locations
}
\keyword{datasets}
