% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{ide_standard}
\alias{ide_standard}
\title{Simulated data from a standard IDE model}
\format{A numeric matrix with 400 rows and 20 columns}
\source{
Randomly generated as part of Easton Huch's MS project at BYU
}
\usage{
ide_standard
}
\description{
A matrix containing simulated data points on a 20 X 20 grid at 20 time points.
The rows correspond to the 400 spatial locations,
and the columns, to the 20 time points.
}
\seealso{
[ide_locations] for a matrix containing the spatial locations
corresponding to the rows of `ide_standard` and
[ide_spatially_varying] for a similar data set generated from
a spatially varying ide model
}
\keyword{datasets}
