% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvnorm}
\alias{mvnorm}
\title{Samples from a multivariate normal distribution}
\usage{
mvnorm(mean, Sigma)
}
\description{
Samples from a multivariate normal distribution
}
\examples{
# Duhh...nothing yet
}
