% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.dstm}
\alias{print.dstm}
\title{Print Method for DSTM Fits}
\usage{
\method{print}{dstm}(x, display = deparse(substitute(x)))
}
\arguments{
\item{x}{A `dstm` object}
}
\description{
Prints a summary for a `dstm` object by calling summary.dstm().
}
\seealso{
summary.dstm
}
