% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.dstm}
\alias{summary.dstm}
\title{Summary Method for DSTM Fits}
\usage{
\method{summary}{dstm}(x, object_name = deparse(substitute(x)))
}
\arguments{
\item{x}{A `dstm` object}

\item{object_name}{The name to be printed in the summary (if desired)}
}
\description{
Prints summary information for `dstm` objects
}
\examples{
num_time_points <- 5
num_spatial_locations <- 100
z <- rnorm(num_time_points * num_spatial_locations)
Y <- matrix(z, nrow=num_spatial_locations, ncol=num_time_points)
rw_model <- dstm_eof(Y, proc_model="RW") # Random walk model
summary(rw_model)

# Identically
print(rw_model)
rw_model
}
