% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dstm}
\alias{dstm}
\title{Fits a dynamic spatio-temporal model (DSTM)}
\usage{
dstm(Y, model = "discount", n_samples = 1L, p = 20L, verbose = FALSE,
  sample_sigma2 = TRUE)
}
\arguments{
\item{Y}{S by T matrix containing response variable at S spatial locations and T time points}

\item{model}{character string; options include `discount`, `sample_G`, `AR`, and `IDE}

\item{n_samples}{integer; number of posterior samples to take}

\item{p}{integer; dimension of G in the state equation \eqn{\theta_{t+1} = G \theta_{t}}}

\item{verbose}{boolean; controls verbosity}

\item{sample_sigma2}{whether boolean; to sample \eqn{\sigma^2}}
}
\description{
Fits a dynamic spatio-temporal model (DSTM)
}
\examples{
# Duhh...nothing yet
}
