% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dstm_IDE}
\alias{dstm_IDE}
\title{Fits a integrodifference equation model (IDE)}
\usage{
dstm_IDE(Y, locs, m_0, C_0, params, n_samples, verbose)
}
\description{
Fits a integrodifference equation model (IDE)
}
\examples{
# Duhh...nothing yet
}
