% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstm.R
\name{dstm_ide}
\alias{dstm_ide}
\title{Fits an integrodifference equation model (IDE)}
\usage{
dstm_ide(Y, locs = NULL, proc_error = "discount", p = 10L,
  n_samples = 1L, sample_sigma2 = TRUE, verbose = FALSE,
  params = NULL)
}
\arguments{
\item{Y}{S by T matrix containing response variable at S spatial locations and T time points}

\item{p}{integer; dimension of G in the state equation \eqn{\theta_{t+1} = G \theta_{t}}}

\item{n_samples}{integer; number of posterior samples to take}

\item{sample_sigma2}{whether boolean; to sample \eqn{\sigma^2}}

\item{verbose}{boolean; controls verbosity}

\item{model}{character string; options include `discount`, `sample_G`, `AR`, and `IDE`}
}
\description{
Fits an integrodifference equation model (IDE)
}
\examples{
# Duhh...nothing yet
}
