% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstm.R
\name{dstm_ide}
\alias{dstm_ide}
\title{Fits an integrodifference equation model (IDE)}
\usage{
dstm_ide(Y, locs = NULL, knot_locs = NULL, proc_error = "IW",
  J = 4L, n_samples = 10L, sample_sigma2 = TRUE, verbose = FALSE,
  params = NULL)
}
\arguments{
\item{Y}{matrix.
S by T matrix containing response variable at S spatial locations and T time points}

\item{locs}{matrix or array.
Spatial locations of observed data. If the locations are the same at each time
point, then `locs` is an S by 2 matrix. 
If the locations change at different time points,
then `locs` is an array with dimension (S, 2, T)}

\item{knot_locs}{integer or matrix.
if integer, then kernel parameters are calculated on a 2-D grid
with dimension (`knot_locs`, `knot_locs`)}

\item{proc_error}{"Discount" or "IW".
If "Discount", then the process error is estimated using a discount factor.
If "IW", then the process error is estimated as a realization from
an inverse-Wishart distribution}

\item{J}{integer. Extent of Fourier approximation.
The size of the state space is 2 * J^2 + 1}

\item{n_samples}{integer; number of posterior samples to draw}

\item{sample_sigma2}{boolean; whether to sample \eqn{\sigma^2}}

\item{verbose}{boolean; verbose = TRUE prints sampling progress}

\item{params}{list; Used to specify hyperparameters}
}
\description{
Detailed description
}
\details{
How to specify hyperparameters such as priors
}
\examples{
# Duhh...nothing yet
}
