% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.dstm}
\alias{summary.dstm}
\title{Summary Method for DSTM Fits}
\usage{
\method{summary}{dstm}(object, object_name = deparse(substitute(object)), ...)
}
\arguments{
\item{object}{A `dstm` object}

\item{object_name}{The name to be printed in the summary (if desired)}

\item{...}{Arguments passed to other methods (necessary for 
S3 generic compatibility)}
}
\description{
Prints summary information for `dstm` objects.
}
\examples{
# Load example data
data("ide_standard", "ide_locations")
mod_ide <- dstm_ide(ide_standard, ide_locations)
summary(mod_ide)
}
\seealso{
[print.summary()]
}
